+++ plugin-dev/Source/Sentry/Private/Windows/WindowsSentrySubsystem.cpp
@@ -13,6 +13,8 @@
 
 void FWindowsSentrySubsystem::InitWithSettings(const USentrySettings* Settings, const FSentryCallbackHandlers& CallbackHandlers)
 {
+	bOutOfProcessScreenshots = Settings->EnableOutOfProcessScreenshots;
+
 	// Detect Wine/Proton before initializing
 	WineProtonInfo = FSentryPlatformDetectionUtils::DetectWineProton();
 
@@ -116,6 +118,15 @@ void FWindowsSentrySubsystem::ConfigureCrashReporterPath(sentry_options_t* Optio
 		return;
 	}
 	sentry_options_set_external_crash_reporter_pathw(Options, *CrashReporterPath);
+}
+
+void FWindowsSentrySubsystem::ConfigureScreenshotCapturing(sentry_options_t* Options)
+{
+	if (bOutOfProcessScreenshots)
+	{
+		UE_LOG(LogSentrySdk, Log, TEXT("Native out-of-process screenshot capturing enabled"));
+		sentry_options_set_attach_screenshot(Options, 1);
+	}
 }
 
 sentry_value_t FWindowsSentrySubsystem::OnCrash(const sentry_ucontext_t* uctx, sentry_value_t event, void* closure)